chore(rearrange-controllers-and-routes): rearrange controllers and routes files - Create menu, orders, users file in both routes and controllers - Populate this with files performing similar functions [Finishes #160980631]
