+++ Server/src/controllers/index.js
@@ -1,20 +1,9 @@
 // Import controllers
-import signUp from './signUp';
-import signIn from './signIn';
-import newOrder from './newOrder';
-import add from './add';
-import allOrders from './allOrders';
-import updateStatus from './updateStatus';
-import specificOrder from './specificOrder';
-import retrieve from './retrieveMenu';
-import history from './history';
-import edit from './edit';
-import remove from './remove';
+import menu from './menu';
+import orders from './orders';
+import users from './users';
 import homePage from './homepage';
 import notFoundPage from './notFoundPage';
 
 // Export controllers
-export {
-  signUp, signIn, newOrder, add, allOrders, updateStatus, specificOrder, retrieve,
-  history, edit, remove, homePage, notFoundPage,
-};
+export { menu, orders, users, homePage, notFoundPage };
