+++ Server/src/index.js
@@ -1,7 +1,7 @@
 // Import modules
 import express from 'express';
 import bodyParser from 'body-parser';
-import apiRoutes from './routes/index';
+import { menu, orders, users, homePage, notFoundPage } from './routes/index';
 
 // Express app
 const app = express();
@@ -13,8 +13,24 @@ const port = process.env.PORT || 3000;
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
-// If the user makes a request to the /api/v1 route, hand control to the apiRoutes
-app.use('/api/v1', apiRoutes);
+// If the user makes a request to the /api/v1/menu route, hand control to the menu route
+app.use('/api/v1/menu', menu);
+
+// If the user makes a request to the /api/v1/orders route, hand control to the orders route
+app.use('/api/v1/orders', orders);
+
+// If the user makes a request to the /api/v1/users route, hand control to the orders route
+app.use('/api/v1/users', orders);
+
+
+// If the user makes a request to the /api/v1/auth route, hand control to the users route
+app.use('/api/v1/auth', users);
+
+// If the user makes a request to the /api/v1/ route, hand control to the homePage route
+app.use('/api/v1', homePage);
+
+// If the user makes a request to the /api/v1/* route, hand control to the notFoundPage route
+app.use('/api/v1/*', notFoundPage);
 
 // Error handler
 app.use((err, req, res, next) => {
